fix: update field

+++ storage/mongo/field.js
@@ -221,15 +221,19 @@ let fieldStore = {
             const Table = mongoConn.models['Table']
             const Field = mongoConn.models['Field']
 
-            if(data.is_system) {
-                throw new Error("This field is system field")
-            }
-
             const fieldBeforUpdate = await Field.findOne(
                 {
                     id: data.id,
                 }
             )
+
+            if(!fieldBeforUpdate) {
+                throw new Error("Field not found")
+            }
+
+            if(fieldBeforUpdate.is_system) {
+                throw new Error("This field is system field")
+            }
             if (con.DYNAMIC_TYPES.includes(data.type) && data.autofill_field && data.autofill_table) {
                 // let autoFillTable = await Table.findOne({
                 //     slug: data.autofill_table,
@@ -536,6 +540,12 @@ let fieldStore = {
             const Field = mongoConn.models['Field']
 
             const deletedField = await Field.findOne({ id: data.id })
+            if(!deletedField) {
+                throw new Error("Field not found")
+            }
+            if(deletedField.is_system) {
+                throw new Error("This field is system field")
+            }
             // const table = await Table.findOne({ id: deletedField.table_id,  })
             const table = await tableVersion(mongoConn, {id: deletedField.table_id}, data.version_id, true)
             const field = await Field.deleteOne({ id: data.id });
